后端用户中心（下）-完成

+++ src/main/java/com/yupi/usercenter1/controller/UserController.java
@@ -10,7 +10,12 @@ import org.springframework.web.bind.annotation.*;
 
 import javax.annotation.Resource;
 import javax.servlet.http.HttpServletRequest;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
+
+import static com.yupi.usercenter1.constant.UserConstant.ADMIN_ROLE;
+import static com.yupi.usercenter1.constant.UserConstant.USER_LOGIN_STATE;
 
 /**
  * ClassName：UserController
@@ -30,48 +35,64 @@ public class UserController {
      */
     @Resource
     private UserService userService;
+
     @PostMapping("/reqister")
-    public Long userRegister(@RequestBody UserRegisterRequest userRegisterRequest){
-        if (userRegisterRequest == null){
+    public Long userRegister(@RequestBody UserRegisterRequest userRegisterRequest) {
+        if (userRegisterRequest == null) {
             return null;
         }
         String userAccount = userRegisterRequest.getUserAccount();
         String userPassword = userRegisterRequest.getUserPassword();
         String checkPassword = userRegisterRequest.getCheckPassword();
-        if (StringUtils.isAnyBlank(userAccount, userPassword, checkPassword)){
+        if (StringUtils.isAnyBlank(userAccount, userPassword, checkPassword)) {
             return null;
         }
         return userService.userRegister(userAccount, userPassword, checkPassword);
     }
 
     @PostMapping("/login")
-    public User userLogin(@RequestBody UserLoginRequest userLoginRequest, HttpServletRequest request){
-        if (userLoginRequest == null){
+    public User userLogin(@RequestBody UserLoginRequest userLoginRequest, HttpServletRequest request) {
+        if (userLoginRequest == null) {
             return null;
         }
         String userAccount = userLoginRequest.getUserAccount();
         String userPassword = userLoginRequest.getUserPassword();
-        if (StringUtils.isAnyBlank(userAccount, userPassword)){
+        if (StringUtils.isAnyBlank(userAccount, userPassword)) {
             return null;
         }
         return userService.userLogin(userAccount, userPassword, request);
     }
 
-   @GetMapping("/search")
-    public List<User> searchUsers(String username){
-       QueryWrapper<User> queryWrapper = new QueryWrapper<>();
-       if (StringUtils.isNotBlank(username)){
-           queryWrapper.like("username", username);
-       }
-       return userService.list(queryWrapper);
-   }
+    @GetMapping("/search")
+    public List<User> searchUsers(String username, HttpServletRequest request) {
+        //仅管理员可查询
+        if (!isAdmin(request)) {
+            return new ArrayList<>();
+        }
+        QueryWrapper<User> queryWrapper = new QueryWrapper<>();
+        if (StringUtils.isNotBlank(username)) {
+            queryWrapper.like("username", username);
+        }
+        List<User> userList = userService.list(queryWrapper);
+        return userList.stream().map(user -> userService.getSafetyUser(user)).collect(Collectors.toList());
+    }
 
-   @PostMapping("/delete")
-    public boolean deleteUser(@RequestBody Long id){
-        if (id <= 0){
+    @PostMapping("/delete")
+    public boolean deleteUser(@RequestBody Long id, HttpServletRequest request) {
+        //仅管理员可删除
+        if (!isAdmin(request)) {
+            return false;
+        }
+        if (id <= 0) {
             return false;
         }
         return userService.removeById(id);
-   }
+    }
 
+    private boolean isAdmin(HttpServletRequest request) {
+        //仅管理员可查询
+        Object userObj = request.getSession().getAttribute(USER_LOGIN_STATE);
+        User user = (User) userObj;
+        return user != null && user.getUserRole() == ADMIN_ROLE;
+    }
 }
